more feedback, ready for merge :)
